// angular_fichajesPi/src/app/intranet/permisos/model/permisoDto.ts
export interface PermisoDto {
  descripcion: string
  diaDesde: string
  diaHasta: string
  horaFinDesde: string
  horaFinHasta: string
  horaInicioDesde: string
  horaInicioHasta: string
  estado: string

  usuarioEmail: string
  usuarioNumero: string
  usuarioNombre: string
  usuarioDni: string
}